requirements.txt
torch>=1.12.0
torch-geometric>=2.3.0
torch-audio>=0.12.0
torch-vision>=0.13.0
pandas>=1.5.0
numpy>=1.21.0,<2.0.0
scikit-learn>=1.1.0
matplotlib==3.8.4
seaborn==0.13.2
scipy>=1.9.0
requests>=2.28.0
tqdm>=4.64.0
optuna>=3.0.0
tensorboard>=2.10.0
pytest>=7.0.0
pyyaml>=6.0
networkx>=2.8.0
plotly>=5.10.0
psutil>=5.9.0
memory-profiler>=0.60.0
holidays>=0.18
